requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
transformers>=4.35.0
supabase>=2.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pydantic>=2.0.0
safetensors>=0.4.0
